Add background/foreground mode  switch and ui for landscape orientation
Changed with gemini-cli with the following prompt (plus some additional instructions for fixes):
"When enabled, Quuppa Tag Android app emits BLE broadcasts that the Quuppa RTLS system uses to compute device location in an indoor space. The app is starting a long running background service but the issue is that Android OS complains about potential battery drain because I've obtained permissions to run the service in the background and because I'm using AlarmManager.setExact. The functionality is broken into this Quuppa Tag Android demo app and the open sourced library at https://github.com/quuppalabs/android-quuppa-taglib that holds the actual service. The 2.0.16-SNAPSHOT version of com.quuppa:android-quuppa-taglib library implements a foreground/background mode (in com.quuppa.tag.QuuppaTag.OperatingMode) to allow user to choose a foreground which doesn't require special permissions to run in the background. The new UI on the main screen OF the app should show to the user whether it's operating in a background mode (i.e. always on) or if the app is only active when it is on the foreground. In landscape orientation, the main UI should have two columns with the background mode toggle on the right and app icon that works as a start/stop button on the left. In portait orientation, the app icon should be on top and the background mode toggle under it. A sample design of the new UI in portrait orientation is in ./quuppa-tag-new-ui-design-foreground-mode.png. Make the necessary changes to the UI and change the value of QuuppaTag.OperatingMode accordingly."